+++ modules/core/dependency/python-ihm/ihm/dumper.py
@@ -92,8 +92,8 @@ class _AuditConformDumper(Dumper):
     def dump(self, system, writer):
         with writer.category("_audit_conform") as lp:
             # Update to match the version of the IHM dictionary we support:
-            lp.write(dict_name="ihm-extension.dic", dict_version="1.22",
-                     dict_location=self.URL % "ac49042")
+            lp.write(dict_name="ihm-extension.dic", dict_version="1.23",
+                     dict_location=self.URL % "8cffaab")
 
 
 class _StructDumper(Dumper):